% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi2021.R
\docType{data}
\name{cdi2021}
\alias{cdi2021}
\title{Taxa de juros CDI - 12}
\format{
Data frame com duas variáveis:
\describe{
\item{dt_cdi}{A data de referencia da taxa.}
\item{pu_ltn010122}{O valor da taxa (em percentual).} }
}
\source{
SGS BCB, serie 12. \url{https://www3.bcb.gov.br/sgspub/}
}
\usage{
cdi2021
}
\description{
Dados referentes ao ano de 2021.
}
\keyword{datasets}
