% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotacoesb3.R
\docType{data}
\name{cotacoesb3}
\alias{cotacoesb3}
\title{Cotação das acoes listadas na B3 - 2021 e 2022}
\format{
Um data frame com 8 variáveis:
\describe{
\item{dtPregao}{O dia do pregao}
\item{codBDI}{Codigo de BDI}
\item{codNegociacao}{Nome da empresa}
\item{tpMercado}{Mercado de negociacao do papel}
\item{nomeResumido}{Nome Resumido da empresa}
\item{especifPapel}{Tipo do papel}
\item{precoUltimo}{Cotacao de fechamento}
}
}
\source{
Adaptacao do arquivo disponivel no site da B3
\url{https://www.b3.com.br/pt_br/market-data-e-indices/servicos-de-dados/market-data/historico/mercado-a-vista/cotacoes-historicas/}
}
\usage{
cotacoesb3
}
\description{
Lista obtida no site da B3. O arquivo completo da B3 possui mais colunas e
um arquivo de layout para entender os itens.
}
\keyword{datasets}
