% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureValue.R
\name{futureValue}
\alias{futureValue}
\title{Valor Futuro
Calcula o valor futuro com a formula convencional de juros compostos}
\usage{
futureValue(pv, i, n)
}
\arguments{
\item{pv}{Valor Presente, em reais}

\item{i}{Taxa de juros, na forma decimal (separador de decimais: ponto)}

\item{n}{Periodo}
}
\value{
Valor Futuro
}
\description{
Valor Futuro
Calcula o valor futuro com a formula convencional de juros compostos
}
\examples{
futureValue(1000, 0.02, 5)
}
