% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltn010122.R
\docType{data}
\name{ltn010122}
\alias{ltn010122}
\title{Tesouro Prefixado com vencimento em 01/01/22}
\format{
Data frame com duas variáveis:
\describe{
\item{dt_ltn}{A data de referencia para o preco unitario.}
\item{pu_ltn010122}{O preco unitario de um titulo com vencimento em 01/01/22.} }
}
\source{
Tesouro Direto. \url{https://www.tesourodireto.com.br/titulos/historico-de-precos-e-taxas.htm}
}
\usage{
ltn010122
}
\description{
Dados referentes ao ano de 2021.
}
\keyword{datasets}
