% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{moda}
\alias{moda}
\title{Moda (valor mais frequente)}
\usage{
moda(vetor)
}
\arguments{
\item{vetor}{Um vetor de numeros.}
}
\value{
Imprime a(s) moda(s).
}
\description{
Moda (valor mais frequente)
}
\examples{
v <- c(1,1,1,2,3,3,3)
moda(v)
}
